bookings-import/requirements.txt
absl-py==0.7.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
